mockups/randomFolder_test.go
package mockups

import (
	"os"
	"testing"
)

func TestCreateDeterministicFolder(t *testing.T) {
	// Test case 1: Create folder with size 0
	t.Run("Create folder with size 0", func(t *testing.T) {
		path := "/tmp"
		size := uint64(0)

		folderPath, err := CreateDeterministicFolder(path, size)
		if err != nil {
			t.Fatalf("Failed to create folder: %v", err)
		}
		defer os.RemoveAll(folderPath)

		// Verify folder exists
		_, err = os.Stat(folderPath)
		if os.IsNotExist(err) {
			t.Errorf("Folder %s does not exist", folderPath)
		}
	})

	// Test case 2: Create folder with non-zero size
	t.Run("Create folder with non-zero size", func(t *testing.T) {
		path := "/tmp"
		size := uint64(1024 * 1024) // 1 MB

		folderPath, err := CreateDeterministicFolder(path, size)
		if err != nil {
			t.Fatalf("Failed to create folder: %v", err)
		}
		defer os.RemoveAll(folderPath)

		// Verify folder exists
		_, err = os.Stat(folderPath)
		if os.IsNotExist(err) {
			t.Errorf("Folder %s does not exist", folderPath)
		}

		// Verify folder size (optional)
		// Add assertions based on your requirements
	})

	// Test case 3: Create nested folders and files
	t.Run("Create nested folders and files", func(t *testing.T) {
		path := "/tmp"
		size := uint64(10 * 1024 * 1024) // 10 MB

		folderPath, err := CreateDeterministicFolder(path, size)
		if err != nil {
			t.Fatalf("Failed to create folder: %v", err)
		}
		defer os.RemoveAll(folderPath)

		// Verify folder exists
		_, err = os.Stat(folderPath)
		if os.IsNotExist(err) {
			t.Errorf("Folder %s does not exist", folderPath)
		}

		// Verify nested folders and files (optional)
		// Add assertions based on your requirements
	})

	// Add more test cases as needed to cover other scenarios
}
